#37 Refactor Minishell Arhitecture [END] (#44)
