+++ includes/Libft/Makefile
@@ -12,7 +12,7 @@ SRCS = ft_atoi.c ft_bzero.c ft_calloc.c ft_isalnum.c ft_isalpha.c ft_isascii.c f
        ft_tolower.c ft_toupper.c ft_strnstr.c \
        get_next_line/get_next_line.c get_next_line/get_next_line_utils.c \
 		ft_lstnew_bonus.c ft_lstadd_front_bonus.c ft_lstsize_bonus.c ft_lstlast_bonus.c ft_lstadd_back_bonus.c ft_lstdelone_bonus.c \
-       ft_lstclear_bonus.c ft_lstiter_bonus.c ft_lstmap_bonus.c 
+       ft_lstclear_bonus.c ft_lstiter_bonus.c ft_lstmap_bonus.c ft_strcpy.c ft_strcat.c ft_is_whitespace.c
 
 OBJS = $(SRCS:.c=.o)
