+++ includes/Libft/ft_strcat.c
@@ -1,36 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   utils.c                                            :+:      :+:    :+:   */
+/*   ft_strcat.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: iatilla- <iatilla-@student.42.fr>          +#+  +:+       +#+        */
+/*   By: schiper <schiper@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/04/23 20:19:31 by iatilla-          #+#    #+#             */
-/*   Updated: 2025/04/23 22:53:30 by iatilla-         ###   ########.fr       */
+/*   Created: 2025/04/24 18:42:14 by schiper           #+#    #+#             */
+/*   Updated: 2025/04/24 18:42:19 by schiper          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "tokener.h"
-
-/**
- * Custom string copy function
- * @param dest: Destination string
- * @param src: Source string
- * @return: Pointer to destination string
- */
-char	*ft_strcpy(char *dest, const char *src)
-{
-	int	i;
-
-	i = 0;
-	while (src[i])
-	{
-		dest[i] = src[i];
-		i++;
-	}
-	dest[i] = '\0';
-	return (dest);
-}
+#include "libft.h"
 
 /**
  * Custom string concatenation function
@@ -54,21 +34,4 @@ char	*ft_strcat(char *dest, const char *src)
 	}
 	dest[i + j] = '\0';
 	return (dest);
-}
-
-/**
- * Free the split array of strings
- * @param dirs: Array of strings to free
- */
-void	free_array(char **dirs)
-{
-	int	i;
-
-	i = 0;
-	while (dirs[i])
-	{
-		free(dirs[i]);
-		i++;
-	}
-	free(dirs);
 }
