+++ sources/input/parser/parse_segment_data.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   token_segment.c                                    :+:      :+:    :+:   */
+/*   parse_segment_data.c                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: iatilla- <iatilla-@student.42.fr>          +#+  +:+       +#+        */
+/*   By: schiper <schiper@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/23 20:19:50 by iatilla-          #+#    #+#             */
-/*   Updated: 2025/04/23 23:58:16 by iatilla-         ###   ########.fr       */
+/*   Updated: 2025/04/24 19:13:26 by schiper          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,14 +28,14 @@ static int	process_operator(t_parse_params *params, int i)
 		|| (params->input[i] == '>' && params->input[i + 1] == '>')
 		|| (params->input[i] == '<' && params->input[i + 1] == '<'))
 	{
-		operator= extract_token(params->input, i, i + 2);
+		operator= extract_string(params->input, i, i + 2);
 		op_type = decide_token_type(operator);
 		add_token(params->tokens, operator, op_type);
 		return (i + 2);
 	}
 	else
 	{
-		operator= extract_token(params->input, i, i + 1);
+		operator= extract_string(params->input, i, i + 1);
 		op_type = decide_token_type(operator);
 		add_token(params->tokens, operator, op_type);
 		return (i + 1);
