+++ sources/input/parser/parser_utils.c
@@ -6,7 +6,7 @@
 /*   By: schiper <schiper@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/23 20:20:48 by iatilla-          #+#    #+#             */
-/*   Updated: 2025/04/23 21:34:28 by schiper          ###   ########.fr       */
+/*   Updated: 2025/04/24 18:44:16 by schiper          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
