+++ sources/memory_management/graph_free.c
