+++ sources/tokener/execution/check_whitespace.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   token_white_space.c                                :+:      :+:    :+:   */
+/*   check_whitespace.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: iatilla- <iatilla-@student.42.fr>          +#+  +:+       +#+        */
+/*   By: schiper <schiper@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/23 20:19:09 by iatilla-          #+#    #+#             */
-/*   Updated: 2025/04/23 23:17:05 by iatilla-         ###   ########.fr       */
+/*   Updated: 2025/04/24 19:02:10 by schiper          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
