+++ sources/tokener/string_utils_for_quotes.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   token_string_utils.c                               :+:      :+:    :+:   */
+/*   string_utils_for_quotes.c                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: iatilla- <iatilla-@student.42.fr>          +#+  +:+       +#+        */
+/*   By: schiper <schiper@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/23 20:19:20 by iatilla-          #+#    #+#             */
-/*   Updated: 2025/04/23 23:16:01 by iatilla-         ###   ########.fr       */
+/*   Updated: 2025/04/24 19:37:25 by schiper          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -81,3 +81,16 @@ int	is_operator(char *input, int i)
 		|| (input[i] == '>' && input[i + 1] == '>') || (input[i] == '<'
 			&& input[i + 1] == '<'));
 }
+
+/**
+ * Outdate Version Maybe ????
+ * Check if a character is an operator
+ * @param c: The character to check
+ * @return: 1 if it's an operator, 0 otherwise
+ */
+int	is_operator_char(char c)
+{
+	return (c == '|' || c == '>' || c == '<' || c == '&' || c == ';');
+}
+
+
